+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Deprecated the endpoint to get a portfolio position in favor of get a holding
+- Deprecated the endpoint to update portfolio position tags in favor of update holding tags
+
 ## 2.159.0 - 2025-05-02
 
 ### Added